+++ apps/embedded-wallet/CHANGELOG.md
@@ -1,5 +1,20 @@
 # embedded-wallet
 
+## 0.4.2
+
+### Patch Changes
+
+- Updated dependencies [1fb5dff]
+  - @idos-network/utils@2.0.1
+  - @idos-network/kwil-infra@2.0.1
+
+## 0.4.1
+
+### Patch Changes
+
+- @idos-network/utils@2.0.0
+- @idos-network/kwil-infra@2.0.0
+
 ## 0.4.0
 
 ### Minor Changes