+++ apps/idos-data-dashboard/CHANGELOG.md
@@ -1,5 +1,29 @@
 # data-dashboard
 
+## 1.4.2
+
+### Patch Changes
+
+- Updated dependencies [1fb5dff]
+  - @idos-network/client@2.0.1
+  - @idos-network/utils@2.0.1
+  - @idos-network/consumer@2.0.1
+  - @idos-network/credentials@2.0.1
+  - @idos-network/issuer@2.0.1
+  - @idos-network/kwil-infra@2.0.1
+
+## 1.4.1
+
+### Patch Changes
+
+- Updated dependencies [8d06f30]
+  - @idos-network/consumer@2.0.0
+  - @idos-network/client@2.0.0
+  - @idos-network/issuer@2.0.0
+  - @idos-network/credentials@2.0.0
+  - @idos-network/utils@2.0.0
+  - @idos-network/kwil-infra@2.0.0
+
 ## 1.4.0
 
 ### Minor Changes