+++ packages/credentials/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @idos-network/credentials
 
+## 2.0.1
+
+### Patch Changes
+
+- Updated dependencies [1fb5dff]
+  - @idos-network/utils@2.0.1
+
+## 2.0.0
+
+### Major Changes
+
+- 8d06f30: New versioned verifiable credentials support
+
+### Patch Changes
+
+- @idos-network/utils@2.0.0
+
 ## 1.5.0
 
 ### Patch Changes